refactor: 기존 detailSettingCardView 새 파일로 분리
