+++ iOS_Wegg/iOS_Wegg/Screen/Schedule/AddSchedule/View/AddScheduleView.swift
@@ -9,11 +9,8 @@ import UIKit
 import Then
 
 protocol AddScheduleGestureDelegate: AnyObject {
-    func didTapCalendarButton()
     func didTapDoneButton()
     func didTapCancelButton()
-    func didSelectStartTime()
-    func didSelectFinishTime()
 }
 
 class AddScheduleView: UIView {
@@ -40,7 +37,7 @@ class AddScheduleView: UIView {
         .notoSans(.medium, size: 16),
         .customGray
     ).then {
-        $0.textAlignment = .left
+        $0.textAlignment = .center
     }
     private lazy var saveLabel = makeLabel(
         "저장",
@@ -105,59 +102,7 @@ class AddScheduleView: UIView {
         , .customGray
     )
 
-    private lazy var detailSettingCardView = makeCardView()
-
-    private lazy var dateLabel = makeLabel("날짜", .notoSans(.medium, size: 16), .customGray)
-    private lazy var calenderImageButton = makeImageButton("calendar")
-    
-    // 날짜 선택 후 보여줄 라벨
-    private lazy var updatedDateLabel = makeLabel(
-        nil,
-        .notoSans(.medium, size: 16),
-        .customGray
-    ).then {
-        $0.isHidden = true
-        $0.textAlignment = .right
-        $0.numberOfLines = 1
-    }
-
-    private lazy var randomVerificationLabel = makeLabel(
-        "랜덤 인증",
-        .notoSans(.medium, size: 16),
-        .customGray
-    )
-
-    private lazy var startTimeButton = makeButton(
-        "00:00",
-        color: .gray1
-    ).then {
-        $0.contentHorizontalAlignment = .right
-    }
-    
-    private lazy var timeRangeLabel = makeLabel(
-        "~",
-        .notoSans(.medium, size: 16),
-        .gray1
-    )
-    
-    private lazy var finishTimeButton = makeButton(
-        "00:00",
-        color: .gray1
-    ).then {
-        $0.contentHorizontalAlignment = .right // 우측 정렬
-    }
-    
-    private lazy var timeRangeStackView = makeStackView(10, .horizontal)
-
-    private lazy var lateAllowanceLabel = makeLabel(
-        "지각 허용",
-        .notoSans(.medium, size: 16),
-        .customGray
-    )
-    private lazy var toggleSwitch = UISwitch().then { $0.onTintColor = .primary }
-
-    private lazy var dividedLine = makeDivider()
-    private lazy var dividedLine2 = makeDivider()
+    private lazy var detailSettingCardView = ScheduleDetailSettingView()
 
     // MARK: - Utility Functions
 
@@ -194,17 +139,6 @@ class AddScheduleView: UIView {
             $0.distribution = .fill
         }
     }
-
-    private func makeButton(
-        _ title: String,
-        color: UIColor
-    ) -> UIButton {
-        return UIButton().then {
-            $0.setTitle(title, for: .normal)
-            $0.setTitleColor(color, for: .normal)
-            $0.titleLabel?.font = .notoSans(.medium, size: 16)
-        }
-    }
     
     private func makeImageButton(
         _ imageName: String
@@ -213,34 +147,14 @@ class AddScheduleView: UIView {
             $0.setImage(UIImage(named: imageName), for: .normal)
         }
     }
-
-    private func makeDivider() -> UIView {
-        return UIView().then {
-            $0.backgroundColor = .customGray2
-            $0.snp.makeConstraints { make in make.height.equalTo(1) }
-        }
-    }
-    
-    private func makeCardView() -> UIView {
-        return UIView().then {
-            $0.backgroundColor = .white
-            $0.layer.borderWidth = 1
-            $0.layer.borderColor = UIColor.secondaryLabel.cgColor
-            $0.layer.cornerRadius = 25
-        }
-    }
     
     // MARK: - Public Functions
     
-    /// 캘린더에서 날짜 선택 후 업데이트하는 함수
-    /// - Parameters:
-    ///     - isHidden: `updatedDateLabel`을 보여줄지 여부
-    ///     - date: 선택된 날짜
-    public func updateDateLabel(isHidden: Bool, date: String?) {
-        updatedDateLabel.isHidden = isHidden
-        if !isHidden {
-            updatedDateLabel.text = date
-        }
+    /// ScheduleDetailSettingView 딜리게이트 설정 함수
+    public func setDetailSettingCardDelegate(
+        _ delegate: ScheduleDetailSettingViewDelegate
+    ) {
+        detailSettingCardView.gestureDelegate = delegate
     }
     
     /// 검색 결과 드롭다운
@@ -260,27 +174,18 @@ class AddScheduleView: UIView {
         return searchResultListView
     }
     
-    /// 랜덤인증 시간 업데이트 함수
-    /// - Parameters:
-    ///     - type: TimePickerType
-    ///     - _ date: 선택된 시간
     public func updateRandomTimeDate(
         type: TimePickertype,
         _ date: String
     ) {
-        switch type {
-        case .startTime:
-            startTimeButton.setTitle(date, for: .normal)
-        case .finishTime:
-            finishTimeButton.setTitle(date, for: .normal)
-        }
+        detailSettingCardView.updateRandomTimeDate(type: type, date)
+    }
+    
+    public func updateDateLabel(isHidden: Bool, date: String?) {
+        detailSettingCardView.updateDateLabel(isHidden: isHidden, date: date)
     }
     
     // MARK: - Action Handler
-    
-    @objc private func handleCalendarButton() {
-        gestureDelegate?.didTapCalendarButton()
-    }
     
     @objc private func handleDoneButton() {
         gestureDelegate?.didTapDoneButton()
@@ -288,14 +193,6 @@ class AddScheduleView: UIView {
     
     @objc private func handleCancelButton() {
         gestureDelegate?.didTapCancelButton()
-    }
-    
-    @objc private func startTimeButtonHandler() {
-        gestureDelegate?.didSelectStartTime()
-    }
-    
-    @objc private func finishTimeButtonHandler() {
-        gestureDelegate?.didSelectFinishTime()
     }
     
     /// 키보드 내리는 핸들러
@@ -313,26 +210,9 @@ private extension AddScheduleView {
         setupGestures()
         addComponents()
         constraints()
-        constraintsDetailSettingView()
     }
     
     func setupGestures() {
-        calenderImageButton.addTarget(
-            self,
-            action: #selector(handleCalendarButton),
-            for: .touchUpInside
-        )
-        startTimeButton.addTarget(
-            self,
-            action: #selector(startTimeButtonHandler),
-            for: .touchUpInside
-        )
-        finishTimeButton.addTarget(
-            self,
-            action: #selector(finishTimeButtonHandler),
-            for: .touchUpInside
-        )
-        
         // 키보드 내리는 제스처 추가
         let dismissKeyboardGesture = UITapGestureRecognizer(
             target: self,
@@ -340,18 +220,10 @@ private extension AddScheduleView {
         )
         dismissKeyboardGesture.cancelsTouchesInView = false
         addGestureRecognizer(dismissKeyboardGesture)
-        
     }
     
     func setupStackView() {
         [cancelLabel, createEggLabel, saveLabel].forEach { headerStackView.addArrangedSubview($0)
-        }
-        [
-            startTimeButton,
-            timeRangeLabel,
-            finishTimeButton
-        ].forEach {
-            timeRangeStackView.addArrangedSubview($0)
         }
     }
     
@@ -368,18 +240,6 @@ private extension AddScheduleView {
             detailSettingCardView
         ].forEach(addSubview)
         bringSubviewToFront(searchResultListView) // 최상위로 배치
-
-        [
-            dateLabel,
-            updatedDateLabel,
-            calenderImageButton,
-            dividedLine,
-            randomVerificationLabel,
-            timeRangeStackView,
-            dividedLine2,
-            lateAllowanceLabel,
-            toggleSwitch
-        ].forEach(detailSettingCardView.addSubview)
     }
     
     func constraints() {
@@ -437,63 +297,5 @@ private extension AddScheduleView {
             make.height.equalTo(174)
         }
         
-    }
-    
-    func constraintsDetailSettingView() {
-        let sideInset: CGFloat = 17
-        let verticalSpacing: CGFloat = 14
-        let labelWidth: CGFloat = 70
-
-        dateLabel.snp.makeConstraints { make in
-            make.top.equalToSuperview().offset(20)
-            make.leading.equalToSuperview().offset(sideInset)
-            make.width.equalTo(labelWidth)
-        }
-        
-        calenderImageButton.snp.makeConstraints { make in
-            make.top.equalTo(dateLabel)
-            make.trailing.equalToSuperview().offset(-sideInset)
-            make.width.height.equalTo(20)
-        }
-        
-        updatedDateLabel.snp.makeConstraints { make in
-            make.top.equalTo(calenderImageButton)
-            make.trailing.lessThanOrEqualTo(calenderImageButton.snp.leading).offset(-5)
-            make.width.equalTo(labelWidth + 70)
-        }
-        
-        dividedLine.snp.makeConstraints { make in
-            make.top.equalTo(dateLabel.snp.bottom).offset(verticalSpacing)
-            make.leading.trailing.equalToSuperview().inset(sideInset)
-        }
-        
-        randomVerificationLabel.snp.makeConstraints { make in
-            make.top.equalTo(dividedLine.snp.bottom).offset(verticalSpacing)
-            make.leading.equalToSuperview().offset(sideInset)
-            make.width.equalTo(labelWidth)
-        }
-        
-        timeRangeStackView.snp.makeConstraints { make in
-            make.top.equalTo(randomVerificationLabel)
-            make.trailing.equalToSuperview().offset(-sideInset)
-            make.width.equalTo(150)
-        }
-        
-        dividedLine2.snp.makeConstraints { make in
-            make.top.equalTo(timeRangeStackView.snp.bottom).offset(verticalSpacing)
-            make.leading.trailing.equalToSuperview().inset(sideInset)
-        }
-        
-        lateAllowanceLabel.snp.makeConstraints { make in
-            make.top.equalTo(dividedLine2.snp.bottom).offset(verticalSpacing)
-            make.leading.equalToSuperview().offset(sideInset)
-            make.bottom.equalToSuperview().offset(-20)
-            make.width.equalTo(labelWidth)
-        }
-        
-        toggleSwitch.snp.makeConstraints { make in
-            make.top.bottom.equalTo(lateAllowanceLabel)
-            make.trailing.equalToSuperview().offset(-sideInset)
-        }
     }
 }
